typecast

+++ 0x1D-search_algorithms/0-linear.c
@@ -10,14 +10,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (array == NULL)
 		return (-1);
 
-	while (i < size)
+	while (i < (int)size)
 	{
-		printf("Value checked array[%zu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 		i++;
